Small check

+++ build_hydra.sh
@@ -153,7 +153,9 @@ clone_repo() {
 	else
 		echo "${PURPLE}Hydra repository already exists${NC}"
 		cd hydra
-		rm -rf build
+		if [ ! -d "build" ]; then
+			rm -rf build
+		fi
 		git pull origin main
 		git submodule update --init --recursive
 	fi
